requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.3
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.5.0
    # via starlette
bcrypt==4.2.0
    # via -r requirements.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
ecdsa==0.19.0
    # via python-jose
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.0
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.5.0
    # via alembic
importlib-resources==6.4.5
    # via alembic
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
mysql-connector-python==9.0.0
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.9.2
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.5.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.12
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.35
    # via
    #   -r requirements.in
    #   alembic
starlette==0.38.6
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   annotated-types
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   uvicorn
urllib3==2.2.3
    # via requests
uvicorn==0.31.0
    # via -r requirements.in
zipp==3.20.2
    # via
    #   importlib-metadata
    #   importlib-resources
